evals/semantic_similarity/requirements.txt
sae_lens==3.21.1
transformer-lens==2.6.0
torch==2.3.0
einops==0.8.0
matplotlib==3.8.4
numpy==1.26.4
pandas==2.1.2
plotly==5.22.0
tqdm==4.66.4
pytest==8.3.2
nbformat==5.10.4
ipykernel==6.29.5
jaxtyping==0.2.25
beartype==0.14.1
ipykernel==6.29.5
seaborn==0.13.2
scikit-learn==1.5.2
sentence-transformers==3.1.1

# Optional utils
nbstripout==0.7.1